sets up authorization middleware and form
