+++ app/Components/Forms/RoomForm.tsx
@@ -55,13 +55,13 @@ export default function RoomForm(){
                     <div className="form-control">
                         <label className="label cursor-pointer">
                             <span className="label-text">Yes</span> 
-                            <input type="radio" name="smoking" value='allowed' className="radio checked:bg-red-500" checked />
+                            <input type="radio" name="smoking" value='allowed' className="radio checked:bg-red-500" defaultChecked />
                         </label>
                         </div>
                         <div className="form-control">
                         <label className="label cursor-pointer">
                             <span className="label-text">no</span> 
-                            <input type="radio" name="smoking" value='not allowed' className="radio checked:bg-blue-500" checked />
+                            <input type="radio" name="smoking" value='not allowed' className="radio checked:bg-blue-500"  />
                         </label>
                     </div>
