+++ app/Components/NavLinks.tsx
@@ -1,0 +1,58 @@
+'use client'
+import Link from 'next/link';
+import { usePathname } from 'next/navigation';
+import clsx from 'clsx'
+
+
+const links = [
+    {name: 'Dasboard', href: '/dashboard'},
+    {name: 'Rooms', href: '/rooms'},
+    {name: 'Roommates', href: '/roommates'},
+]
+
+export default function NavLinks() {
+    const pathname = usePathname();
+    return (
+    <>
+      <div className="navbar container bg-slate-400">            
+            <div className="flex-none">
+                <ul className="menu menu-horizontal px-1">
+                {links.map((link) => {
+                    return (
+                    <Link
+                        key={link.name}
+                        href={link.href}
+                        className={clsx(
+                        'flex h-[48px] grow items-center justify-center gap-2 rounded-md bg-gray-50 p-3 text-sm font-medium hover:bg-sky-100 hover:text-blue-600 md:flex-none md:justify-start md:p-2 md:px-3',
+                        {
+                            'bg-sky-100 text-blue-600': pathname === link.href,
+                        },
+                        )}
+                    >
+                        <p className="hidden md:block">{link.name}</p>
+                    </Link>
+                    );
+                })}
+                {/* <li>
+                    <details>
+                    <summary>
+                        Parent
+                    </summary>
+                    <ul className="p-2 bg-base-100 rounded-t-none">
+                        <li><a>Link 1</a></li>
+                        <li><a>Link 2</a></li>
+                    </ul>
+                    </details>
+                </li> */}
+                </ul>
+
+                
+                
+            </div>
+             
+        </div>
+    
+    </>
+
+    )
+}
