+++ app/lib/actions/UserActions/actions.ts
@@ -58,6 +58,7 @@ export async function authenticate(
                 return 'Something went wrong.';
             }
         }
+        throw error;
     }
 }
