+++ auth.config.ts
@@ -8,7 +8,11 @@ export const authConfig = {
     authorized({ auth, request: { nextUrl } }) {
       const isLoggedIn = !!auth?.user;
       const isOnDashboard = nextUrl.pathname.startsWith('/dashboard');
-      if (isOnDashboard) {
+      const isOnRooms = nextUrl.pathname.startsWith('/rooms');
+      const isOnRoommates = nextUrl.pathname.startsWith('/roommates');
+
+
+      if (isOnDashboard || isOnRooms || isOnRoommates) {
         if (isLoggedIn) return true;
         return false; // Redirect unauthenticated users to login page
       } else if (isLoggedIn) {
