+++ package.json
@@ -18,7 +18,7 @@
     "bcrypt": "^5.1.1",
     "clsx": "^2.0.0",
     "next": "^14.0.2",
-    "next-auth": "^5.0.0-beta.5",
+    "next-auth": "^5.0.0-beta.4",
     "postcss": "8.4.31",
     "react": "18.2.0",
     "react-dom": "18.2.0",
